+++ tlias-web-management/src/main/java/com/sheng/service/impl/DeptServiceImpl.java
@@ -1,6 +1,7 @@
 package com.sheng.service.impl;
 
 import com.sheng.mapper.DeptMapper;
+import com.sheng.mapper.EmpMapper;
 import com.sheng.pojo.Dept;
 import com.sheng.service.DeptService;
 import org.springframework.stereotype.Service;
@@ -10,9 +11,12 @@ import java.util.List;
 @Service
 public class DeptServiceImpl implements DeptService {
 	private final DeptMapper deptMapper;
+	private final EmpMapper empMapper;
 
-	public DeptServiceImpl(DeptMapper deptMapper) {
+	public DeptServiceImpl(DeptMapper deptMapper, EmpMapper empMapper) {
+
 		this.deptMapper = deptMapper;
+		this.empMapper = empMapper;
 	}
 
 	@Override
@@ -33,7 +37,11 @@ public class DeptServiceImpl implements DeptService {
 
 	@Override
 	public void deleteDeptById(Integer id) {
-		deptMapper.deleteDeptById(id);
+		if (!empMapper.findEmpByDeptId(id).isEmpty()) {
+			throw new RuntimeException("该部门下有员工，不能删除");
+		} else {
+			deptMapper.deleteDeptById(id);
+		}
 	}
 
 	@Override
